requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
flake8==3.8.3
gunicorn==20.0.4
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
